renamed chat_thread.py to chat_model_openai.py
