+++ src/input_triggers/input_triggers_main.py
@@ -15,7 +15,7 @@ import pathlib
 
 # Now imports relative to src should work if sys.path is configured correctly
 from input_triggers.input_triggers import InputTrigger
-from ras.chat_thread import get_gpt_handler
+from ras.chat_model_openai import get_gpt_handler
 
 # Use logging instead of print for better control
 logger = logging.getLogger(__name__)
