+++ src/ras/chat_model_openai.py
